pyproject.toml
[project]
name = "kuckuck"
description = "Lokale Pseudonymisierung personenbezogener Daten in Textdateien vor der Verarbeitung durch Cloud-LLMs"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Hochfrequenz Unternehmensberatung GmbH", email = "info+github@hochfrequenz.de" }]
keywords = ["pseudonymization", "pii", "gdpr", "dsgvo", "llm", "privacy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security",
    "Topic :: Text Processing",
]
dependencies = [
    "pydantic>=2.11",
    "pydantic-settings>=2.9",
    "python-dotenv>=1.1",
    "cryptography>=44",
    "phonenumbers>=8.13",
    "pyahocorasick>=2.1",
    "email-validator>=2.2",
    # Format-aware preprocessors. Lower bounds chosen to track the
    # minimum versions actually exercised by the test suite. We avoid
    # upper bounds to stay flexible for dependency-update bots; the
    # surface we use is small and stable in each library.
    "extract-msg>=0.52",
    "markdown-it-py>=3",
    "lxml>=5",
    "selectolax>=0.3",
]
dynamic = ["readme", "version"]

[project.optional-dependencies]
cli = [
    "typer>=0.15"
]
tests = [
    "pytest==9.0.3",
    "syrupy==5.1.0",
    "hypothesis==6.152.1"
]
linting = [
    "pylint==4.0.5",
    "pylint-pydantic==0.4.1"
]
type_check = [
    "mypy==1.20.2"
]
spell_check = [
    "codespell==2.4.2"
]
coverage = [
    "coverage==7.13.5"
]
formatting = [
    "black==26.3.1",
    "isort==8.0.1"
]
packaging = [
    "build==1.4.3",
    "twine==6.2.0"
]
build_executable = [
    "pyinstaller==6.19.0"
]
dev = [
    "pip-tools"
]


[project.urls]
Changelog = "https://github.com/Hochfrequenz/kuckuck/releases"
Homepage = "https://github.com/Hochfrequenz/kuckuck"

[project.scripts]
kuckuck = "kuckuck.__main__:main"

[tool.black]
line-length = 120
target_version = ["py311", "py312", "py313", "py314"]

[tool.isort]
line_length = 120
profile = "black"

[tool.pylint.MASTER]
load-plugins = "pylint_pydantic"

[tool.pylint."MESSAGES CONTROL"]
max-line-length = 120
disable = [
    "too-few-public-methods",   # detectors implement a single-method Protocol
    "duplicate-code",           # three detectors build Span objects identically by design
    "c-extension-no-member",    # pyahocorasick is a C extension
    "no-member",                # pylint-pydantic handles most FieldInfo cases; a few slip through
]

[tool.pytest.ini_options]
markers = ["snapshot: mark a test as a snapshot test"]

[tool.mypy]
disable_error_code = []

[build-system]
requires = ["hatchling>=1.29.0", "hatch-vcs", "hatch-fancy-pypi-readme"]
build-backend = "hatchling.build"

[tool.hatch.metadata.hooks.fancy-pypi-readme]
content-type = "text/markdown"
fragments = [{ path = "README.md" }]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/_kuckuck_version.py"
template = '''
version = "{version}"
'''

[tool.hatch.build.targets.sdist]
exclude = ["/unittests"]

[tool.hatch.build.targets.wheel]
only-include = ["src"]
sources = ["src"]
